Missed )s

+++ python/test.py
@@ -7,8 +7,8 @@ address = 0x80
 clicks = 1000
 click_vel = 150
 ACCELERATION = 300
-rc.SpeedAccelDistanceM1M2(address=address,accel=ACCELERATION,speed1=int(click_vel),distance1=int(abs(clicks/2)),speed2=int(click_vel),distance2=int(abs(clicks/2)),buffer=1
-rc.SpeedAccelDistanceM1M2(address=address,accel=ACCELERATION,speed1=0,distance1=int(abs(clicks/2)),speed2=0,distance2=int(abs(clicks/2)),buffer=0
+rc.SpeedAccelDistanceM1M2(address=address,accel=ACCELERATION,speed1=click_vel,distance1=int(abs(clicks/2)),speed2=int(click_vel),distance2=int(abs(clicks/2)),buffer=1)
+rc.SpeedAccelDistanceM1M2(address=address,accel=ACCELERATION,speed1=0,distance1=int(abs(clicks/2)),speed2=0,distance2=int(abs(clicks/2)),buffer=0)
 buffers = (0,0,0)
 while (buffers[1]!=0x80 and buffers[2]!=0x80):
     buffers = rc.ReadBuffers(address);
